#CDE-24 Move the components to common repo
